requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.3
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.3.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
twilio==9.0.3
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
